+++ integration_tests/test_suites/k8s-test-suite/tests/test_external_asset.py
@@ -4,13 +4,13 @@ from contextlib import contextmanager
 import kubernetes
 import pytest
 from dagster import AssetExecutionContext, asset, materialize
-from dagster._core.ext.client import (
+from dagster._core.pipes.client import (
     ExtContextInjector,
 )
-from dagster._core.ext.utils import ExtEnvContextInjector, ext_protocol
+from dagster._core.pipes.utils import ExtEnvContextInjector, ext_protocol
 from dagster_k8s import execute_k8s_job
 from dagster_k8s.client import DagsterKubernetesClient
-from dagster_k8s.ext import ExtK8sPod, K8sPodLogsMessageReader
+from dagster_k8s.pipes import ExtK8sPod, K8sPodLogsMessageReader
 from dagster_pipes import ExtContextData, ExtDefaultContextLoader
 from dagster_test.test_project import (
     get_test_project_docker_image,
