+++ python_modules/dagster/dagster/_core/pipes/subprocess.py
@@ -7,13 +7,13 @@ from dagster import _check as check
 from dagster._core.definitions.resource_annotation import ResourceParam
 from dagster._core.errors import DagsterExternalExecutionError
 from dagster._core.execution.context.compute import OpExecutionContext
-from dagster._core.ext.client import (
+from dagster._core.pipes.client import (
     ExtClient,
     ExtContextInjector,
     ExtMessageReader,
 )
-from dagster._core.ext.context import ExtResult
-from dagster._core.ext.utils import (
+from dagster._core.pipes.context import ExtResult
+from dagster._core.pipes.utils import (
     ExtTempFileContextInjector,
     ExtTempFileMessageReader,
     ext_protocol,
