+++ python_modules/libraries/dagster-aws/dagster_aws/pipes.py
@@ -6,10 +6,10 @@ from typing import Iterator, Optional
 import boto3
 import dagster._check as check
 from botocore.exceptions import ClientError
-from dagster._core.ext.client import (
+from dagster._core.pipes.client import (
     ExtParams,
 )
-from dagster._core.ext.utils import ExtBlobStoreMessageReader
+from dagster._core.pipes.utils import ExtBlobStoreMessageReader
 
 
 class ExtS3MessageReader(ExtBlobStoreMessageReader):
